+++ org/bovigo/vfs/visitor/vfsStreamStructureVisitor.php
@@ -2,14 +2,14 @@
 
 namespace org\bovigo\vfs\visitor;
 
-use bovigo\vfs\visitor\vfsStreamStructureVisitor as Base;
+use bovigo\vfs\visitor\StructureInspector as Base;
 
-class_exists('bovigo\vfs\visitor\vfsStreamStructureVisitor');
+class_exists('bovigo\vfs\visitor\StructureInspector');
 
-@trigger_error('Using the "org\bovigo\vfs\visitor\vfsStreamStructureVisitor" class is deprecated since version 1.7 and will be removed in version 2, use "bovigo\vfs\visitor\vfsStreamStructureVisitor" instead.', E_USER_DEPRECATED);
+@trigger_error('Using the "org\bovigo\vfs\visitor\vfsStreamStructureVisitor" class is deprecated since version 1.7 and will be removed in version 2, use "bovigo\vfs\visitor\StructureInspector" instead.', E_USER_DEPRECATED);
 
 if (\false) {
-    /** @deprecated since 1.7, use "bovigo\vfs\vfsStreamStructureVisitor" instead */
+    /** @deprecated since 1.7, use "bovigo\vfs\StructureInspector" instead */
     class vfsStreamStructureVisitor extends Base
     {
     }